% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{alchemer_db_path}
\alias{alchemer_db_path}
\title{Alchemer application database directory}
\usage{
alchemer_db_path(db = NULL)
}
\arguments{
\item{db}{Directory override. Defaults to \code{Sys.getenv("ALCHEMER_DB")}.}
}
\value{
A single string path.
}
\description{
Alchemer application database directory
}
\keyword{internal}
