% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alchemeR-package.R
\docType{package}
\name{alchemeR-package}
\alias{alchemeR}
\alias{alchemeR-package}
\title{alchemeR: Download 'Alchemer' Survey Data}
\description{
Provides functions to access Alchemer data directly into R using the 'Alchemer' API. 'Alchemer' \url{https://www.alchemer.com/} is an online survey and data collection software platform. See \url{https://apihelp.alchemer.com/help} for more information about their API. This package is community-maintained and is not officially supported by 'Alchemer'.
}
\author{
\strong{Maintainer}: Rousell Greg \email{grousell@gmail.com}

}
\keyword{internal}
