% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data_dictionary.R
\name{fetch_data_dictionary}
\alias{fetch_data_dictionary}
\title{Fetch Alchemer Data Dictionary}
\usage{
fetch_data_dictionary(survey_id, token, secret_key)
}
\arguments{
\item{survey_id}{Alchemer survey number, retrieved from survey URL}

\item{token}{Alchemer API token (https://apihelp.alchemer.com/help/authentication)}

\item{secret_key}{Alchemer API secret key}
}
\value{
A data frame of question numbers and question labels
}
\description{
A function that retrieves a data dictionary from an Alchemer survey.
Requires user to have set up an API with Alchemer, including
a token and secret key. If custom reporting values are used in Alchemer, then
the function will download the custome value.
}
\examples{
# fetch_data_dictionary("SURVEY_ID","YOUR-ALCHEMER_TOKEN","YOUR-ALCHEMER_SECRET_KEY")
}
