% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{fetch_data_dictionary}
\alias{fetch_data_dictionary}
\title{Fetch an Alchemer survey's data dictionary}
\usage{
fetch_data_dictionary(survey_id, token = "token", secret_key = "secret_key")
}
\arguments{
\item{survey_id}{Alchemer survey number, retrieved from the survey URL.}

\item{token}{Alchemer API token.}

\item{secret_key}{Alchemer API secret key.}
}
\value{
A tibble of question metadata.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Superseded by \code{\link[=alchemer_questions]{alchemer_questions()}}.
}
