% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_survey.R
\name{fetch_survey}
\alias{fetch_survey}
\title{Fetch Alchemer Survey Data
This function retrieves survey data from an Alchemer account and saves a .csv file in a project folder.
Requires user to have set up an API with Alchemer, including a token and secret key}
\usage{
fetch_survey(survey_id, token, secret_key)
}
\arguments{
\item{survey_id}{Alchemer survey number, retrieved from survey URL}

\item{token}{Alchemer API token (https://apihelp.alchemer.com/help/authentication)}

\item{secret_key}{Alchemer API secret key}
}
\value{
Downloaded file of survey responses
}
\description{
Fetch Alchemer Survey Data
This function retrieves survey data from an Alchemer account and saves a .csv file in a project folder.
Requires user to have set up an API with Alchemer, including a token and secret key
}
\examples{
# alchemer_survey("SURVEY_ID", <YOUR-ALCHEMER_TOKEN>, <YOUR-ALCHEMER_SECRET_KEY> )
}
